proj2 -> tests should convert to ints, otherwise 150 is not the same as 150.0

+++ projects/project02/tests/q1.py
@@ -6,7 +6,7 @@ test = {
       'cases': [
         {
           'code': r"""
-          >>> round(scenario_1_months_rate, 0)
+          >>> int(round(scenario_1_months_rate, 0))
           150
           """,
           'hidden': False,
